chore: refactor error pages
