+++ apps/recnet/src/components/rec/RecForm.tsx
@@ -13,9 +13,9 @@ import { toast } from "sonner";
 import { z } from "zod";
 
 import { trpc } from "@recnet/recnet-web/app/_trpc/client";
-import { ReportEmailAccount } from "@recnet/recnet-web/app/not-found";
 import { Accordion } from "@recnet/recnet-web/components/Accordion";
 import { RecNetLink } from "@recnet/recnet-web/components/Link";
+import { ReportEmailAccount } from "@recnet/recnet-web/components/error";
 import { RecArticleForm } from "@recnet/recnet-web/components/rec/RecArticleForm";
 import { cn } from "@recnet/recnet-web/utils/cn";
